requirements.txt
psycopg2-binary 
python-dotenv 
openai
langchain
beautifulsoup4
flask
apscheduler